repositorio de sistema medico
Se agrego la pre-visualizacion de imagenes de perfil.
y el selector de consultorio de doctor al modificar una ficha ya muetra el dato alamacenado.

Se habilito la opcion de status para deshabilitar a los a los trabajadores o empleados.

Se tradujeron los mensajes de validacion al español, se corrigieron varios errores

se agregaron a las tablas de todos los formularios la indicacion de activado o desactivado

Al momento de eliminar a un usuario aparece un mensaje de confirmacion

Se corrigieron los probelmas con las pestaña pso y la crearcion y edicion.

Se agrego al login

+++ routes/web.php
@@ -2,6 +2,8 @@
 
 use Illuminate\Support\Facades\Route;
 /*--------------------------------------------------------------------------------------------------------------------*/
+use App\Http\Controllers\Seguridad\SeguridadController;
+/*--------------------------------------------------------------------------------------------------------------------*/
 use App\Http\Controllers\Paginas_principal\PaginasPrincipalController;
 /*--------------------------------------------------------------------------------------------------------------------*/
 use App\Http\Controllers\Administrador\AdministradoresController;
@@ -22,34 +24,37 @@ use App\Http\Controllers\Secretaria\PacienteSecretariaController;
 |
 */
 
-Route::get('/', function () {
+/*Route::get('/', function () {
     return view('welcome');
-});
+});*/
 
-
+/*---------------------------------------Tutas de seguridad-----------------------------------------------------*/
+Route::get('login', [SeguridadController::class,'vistaLogin'])->name('login');
+Route::post('login', [SeguridadController::class,'loginIngresar'])->name('login');
+Route::post('logout', [SeguridadController::class,'logout'])->name('logout');
 /*---------------------------------------Ruta index Administrador-----------------------------------------------------*/
-Route::get('administrador/post', [PaginasPrincipalController::class,'lista_post'])->name('administrador.lista.post');
-Route::get('administrador/post/crear', [PaginasPrincipalController::class,'crear_post'])->name('administrador.crear.post');
-Route::post('administrador/post', [PaginasPrincipalController::class,'guardar_post'])->name('administrador.guardar.post');
-Route::get('administrador/post/{id}/editar', [PaginasPrincipalController::class,'editar_post'])->name('administrador.editar.post');
-Route::put('administrador/post/{id}', [PaginasPrincipalController::class,'actualizar_post'])->name('administrador.actualizar.post');
-Route::delete('administrador/post/{id}', [PaginasPrincipalController::class,'eliminar_post'])->name('administrador.elimianr.post');
+Route::get('administrador/post', [PaginasPrincipalController::class,'lista_post'])->name('administrador.lista.post')->middleware('auth');
+Route::get('administrador/post/crear', [PaginasPrincipalController::class,'crear_post'])->name('administrador.crear.post')->middleware('auth');
+Route::post('administrador/post', [PaginasPrincipalController::class,'guardar_post'])->name('administrador.guardar.post')->middleware('auth');
+Route::get('administrador/post/{id}/editar', [PaginasPrincipalController::class,'editar_post'])->name('administrador.editar.post')->middleware('auth');
+Route::put('administrador/post/{id}', [PaginasPrincipalController::class,'actualizar_post'])->name('administrador.actualizar.post')->middleware('auth');
+Route::delete('administrador/post/{id}', [PaginasPrincipalController::class,'eliminar_post'])->name('administrador.elimianr.post')->middleware('auth');
 /*--------------------------------------------------------------------------------------------------------------------*/
 
-Route::resource('administrador/admin', AdministradoresController::class);
-Route::resource('administrador/doctor', DoctorAdminController::class);
-Route::resource('administrador/secretaria', SecretariaAdminController::class);
-Route::resource('administrador/paciente', PacienteAdminController::class);
+Route::resource('administrador/admin', AdministradoresController::class)->middleware('auth');
+Route::resource('administrador/doctor', DoctorAdminController::class)->middleware('auth');
+Route::resource('administrador/secretaria', SecretariaAdminController::class)->middleware('auth');
+Route::resource('administrador/paciente', PacienteAdminController::class)->middleware('auth');
 /*--------------------------------------------------------------------------------------------------------------------*/
-Route::get('Secretaria/doctor', [DoctorSecretariaController::class, 'index'])->name('secretaria.doctor.index');
-Route::get('Secretaria/doctor/crear', [DoctorSecretariaController::class, 'create'])->name('secretaria.doctor.create');
-Route::post('Secretaria/doctor', [DoctorSecretariaController::class, 'store'])->name('secretaria.doctor.store');
-Route::get('Secretaria/doctor/{id}/editar', [DoctorSecretariaController::class, 'edit'])->name('secretaria.doctor.edit');
-Route::put('Secretaria/doctor/{id}', [DoctorSecretariaController::class, 'update'])->name('secretaria.doctor.update');
+Route::get('Secretaria/doctor', [DoctorSecretariaController::class, 'index'])->name('secretaria.doctor.index')->middleware('auth');
+Route::get('Secretaria/doctor/crear', [DoctorSecretariaController::class, 'create'])->name('secretaria.doctor.create')->middleware('auth');
+Route::post('Secretaria/doctor', [DoctorSecretariaController::class, 'store'])->name('secretaria.doctor.store')->middleware('auth');
+Route::get('Secretaria/doctor/{id}/editar', [DoctorSecretariaController::class, 'edit'])->name('secretaria.doctor.edit')->middleware('auth');
+Route::put('Secretaria/doctor/{id}', [DoctorSecretariaController::class, 'update'])->name('secretaria.doctor.update')->middleware('auth');
 /*--------------------------------------------------------------------------------------------------------------------*/
-Route::get('secretaria/paciente', [PacienteSecretariaController::class, 'index'])->name('secretaria.paciente.index');
-Route::get('secretaria/paciente/crear', [PacienteSecretariaController::class, 'create'])->name('secretaria.paciente.create');
-Route::post('secretaria/paciente', [PacienteSecretariaController::class, 'store'])->name('secretaria.paciente.store');
-Route::get('secretaria/paciente/{id}/editar', [PacienteSecretariaController::class, 'edit'])->name('secretaria.paciente.edit');
-Route::put('secretaria/paciente/{id}', [PacienteSecretariaController::class, 'update'])->name('secretaria.paciente.update');
+Route::get('secretaria/paciente', [PacienteSecretariaController::class, 'index'])->name('secretaria.paciente.index')->middleware('auth');
+Route::get('secretaria/paciente/crear', [PacienteSecretariaController::class, 'create'])->name('secretaria.paciente.create')->middleware('auth');
+Route::post('secretaria/paciente', [PacienteSecretariaController::class, 'store'])->name('secretaria.paciente.store')->middleware('auth');
+Route::get('secretaria/paciente/{id}/editar', [PacienteSecretariaController::class, 'edit'])->name('secretaria.paciente.edit')->middleware('auth');
+Route::put('secretaria/paciente/{id}', [PacienteSecretariaController::class, 'update'])->name('secretaria.paciente.update')->middleware('auth');
 /*--------------------------------------------------------------------------------------------------------------------*/
